% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_hcd.R
\name{collapse_hcd}
\alias{collapse_hcd}
\title{Bind a list of HCD data sets, row-wise.}
\usage{
collapse_hcd(l)
}
\arguments{
\item{l}{a list whose elements are tibbles that should be bound, row-wise.}
}
\value{
A \code{\link{tbl_df}} of the data after row binding each element of input.
}
\description{
Given a list of HCD data sets, \code{collapse_hcd} binds together each element of the list (HCD dataset) row-wise into a single tibble.
}
\author{
Gavin L. Simpson
}
