% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_proximity.R
\name{find_proximity}
\alias{find_proximity}
\title{Find a Climate Data station by proximity to a target}
\usage{
find_proximity(target, mindist = 0, maxdist = 100, sort = TRUE)
}
\arguments{
\item{target}{numeric; number of the target (reference) station}

\item{mindist}{numeric; minimum distance from the target station (defaults to 0)}

\item{maxdist}{numeric; maximum distance from the target station (defaults to 100)}

\item{sort}{Boolean; if set to TRUE (default), will sort the resultant table by distance from `target`}
}
\value{
An object of class \code{"hcd_station_list"}, which is a \code{"tbl_df"}, containing details of any matching HCD stations.
}
\description{
Search for stations in the Historical Climate Data inventory by proximity to a target station.
}
\examples{
find_proximity(5051)
}
\author{
Conor I. Anderson
}

