% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcd_daily.R
\name{hcd_daily}
\alias{hcd_daily}
\title{Download daily Historical Climate Data records}
\usage{
hcd_daily(station, year, collapse = TRUE, progress = TRUE,
  cache = FALSE, ...)
}
\arguments{
\item{station}{Character or numeric; one or more HCD station IDs.}

\item{year}{numeric; vector of one or more years}

\item{collapse}{Logical; if \code{TRUE} and multiple \code{station}s are requested a single \code{data frame} is returned, formed by row-binding the data for each station and prepending a station identifier variable.}

\item{progress}{Logical; if \code{TRUE}, a bar is shown indicating progress in downloading station data from the HCD website.}

\item{cache}{Logical; if \code{TRUE}, the processed data will be cached in your user cache dir for future use.}

\item{...}{Further arguments passed to \code{\link{read_hcd}}.}
}
\value{
A \code{\link{tbl_df}} containing the requested monthly climate data
}
\description{
Daily Historical Climate Data for stated station IDs are downloaded from the Government of Canada's Historical Climate Data website.
}
\author{
Gavin L. Simpson
}
