% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcd_key.R
\name{hcd_key}
\alias{hcd_key}
\title{Generate keys to Historical Climate Data records}
\usage{
hcd_key(station, timescale = c("monthly", "daily", "hourly"), year,
  month)
}
\arguments{
\item{station}{character; one or more HCD station IDs.}

\item{timescale}{character; the timescale of the requested Historical Climate Data records.}

\item{year}{numeric; vector of one or more years. Needed for daily or hourly data only.}

\item{month}{numeric; vector of one or months. Needed for hourly data only.}
}
\value{
A character vector containing keys for the requested stations and time scales.
}
\description{
Form keys to requested Canadian Historical Climate Data records.
}
\examples{
keys <- hcd_key(c(2855, 1707), timescale = "monthly")
keys
head(keys$key)
}
\author{
Gavin L. Simpson
}
