% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcd_monthly.R
\name{hcd_monthly}
\alias{hcd_monthly}
\title{Download monthly Historical Climate Data records}
\usage{
hcd_monthly(station, collapse = TRUE, ...)
}
\arguments{
\item{station}{Character; one or more (latter not currently implemented) HCD station IDs.}

\item{collapse}{Logical; if multiple \code{station}s are requested a single \code{data_frame} is returned, formed by row-binding the data for each station and prepending a station identifier variable.}

\item{...}{Further arguments passed to \code{\link{read_hcd}}.}
}
\value{
A \code{\link{tbl_df}} containing the requested monthly climate data
}
\description{
Monthly Historical Climate Data for stated station IDs are downloaded from the Government of Canada's Historical Climate Data website.
}
\author{
Gavin L. Simpson
}

