% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcd_url.R
\name{hcd_url}
\alias{hcd_url}
\title{Generate URLs to Historical Climate Data records}
\usage{
hcd_url(station, timescale = c("monthly", "daily", "hourly"), year, month)
}
\arguments{
\item{station}{character; one or more HCD station IDs.}

\item{timescale}{character; the timescale of the requested Historical Climate Data records.}

\item{year}{numeric; vector of one or more years. Needed for daily or hourly data only.}

\item{month}{numeric; vector of one or months. Needed for hourly data only.}
}
\value{
A character vector containing URLs for the requested stations and time scales.
}
\description{
Form URLs to requested Canadian Historical Climate Data records.
}
\examples{
urls <- hcd_url(c(2855, 1707), timescale = "monthly")
urls
head(urls$url)
}
\author{
Gavin L. Simpson
}
