% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hcd.R
\name{read_hcd}
\alias{read_hcd}
\title{Read Historical Climate Data files}
\usage{
read_hcd(file, flags = FALSE, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection. See argument \code{file} of \code{\link{read_csv}}.}

\item{flags}{Logical; should variable flag columns be preserved. The default results in these columns being dropped from the returned data.}

\item{...}{Further arguments passed to \code{\link{read_csv}}.}
}
\value{
A \code{\link{tbl_df}}.
}
\description{
Reads data from CSV files download from the Canada Historical Climate Data website. The function knows about hourly, daily, and monthly data formatted files.
}
\examples{
read_hcd(system.file("extdata/2855-monthly-data.csv", package = "canadaHCD"))
}
\author{
Gavin L. Simpson
}

