% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_id.R
\name{station_id}
\alias{station_id}
\title{Returns the StationID for a supplied ClimateID}
\usage{
station_id(id)
}
\arguments{
\item{id}{numeric or character; the ClimateIDs whose StationIDs are to be
returned.}
}
\description{
Returns the StationID for a supplied ClimateID
}
