% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_station_data.R
\name{update_station_data}
\alias{update_station_data}
\title{Download station data}
\usage{
update_station_data(force = FALSE)
}
\arguments{
\item{force}{logical; should we force the station data to be updated?}
}
\description{
Download station data
}
\keyword{internal}
