% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_id.R
\name{climate_id}
\alias{climate_id}
\title{Returns the ClimateID for a supplied StationID}
\usage{
climate_id(station)
}
\arguments{
\item{station}{numeric or character; the StationIDs whose ClimateIDs are to
be returned.}
}
\value{
A character vector of ClimateIDs.
}
\description{
Returns the ClimateID for a supplied StationID
}
