% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcd_download.R
\name{hcd_download}
\alias{hcd_download}
\title{Download CSV files of Historical Climate Data records}
\usage{
hcd_download(
  station,
  file,
  dir = tempdir(),
  timescale = c("monthly", "daily", "hourly"),
  year = NULL,
  month = NULL,
  progress = TRUE
)
}
\arguments{
\item{station}{character; one or more HCD station IDs.}

\item{file}{character; one or more file names. If not supplied, file names of the form \code{station-year-month.csv}, with \code{year} and \code{month} appended as determined by argument \code{timescale}.}

\item{dir}{character; vector of one or more directory paths in which CSV files will be saved.scale of the requested Historical Climate Data records. If multiple paths supplied, there should be one path per station.}

\item{timescale}{character; the timescale of the requested Historical Climate Data records.}

\item{year}{numeric; vector of one or more years. Needed for daily or hourly data only.}

\item{month}{numeric; vector of one or months. Needed for hourly data only.}

\item{progress}{logical; if \code{TRUE}, a bar is shown indicating progress in downloading station data from the HCD website.}
}
\value{
A character vector containing file paths to the downloaded CSVs is returned invisibly.
}
\description{
Downloads raw CSV data files from the Canadian Historical Climate Data website. These are the raw files you would obtain if you accessed a data record through the HCD website.
}
\examples{
f <- hcd_download(1707, timescale = "monthly")
file.exists(f)

## specifying multiple storage folders
dirs <- file.path(tempdir(), c("2855", "1707"))
dir.create(dirs[1])
dir.create(dirs[2])
f <- hcd_download(c(2855, 1707), timescale = "monthly", dir = dirs)
file.exists(f)
## ...and daily
f <- hcd_download(c(2855, 1707), timescale = "daily", dir = dirs,
                  year = 1999:2000)
file.exists(f)
}
\author{
Gavin L. Simpson
}
